\name{print.PrenticeGloeckler.test}
\alias{print.PrenticeGloeckler.test}
\title{Print Regression for Grouped Survival Data Function Object}
\description{
Print coefficient returned by PrenticeGloeckler function.
}
\usage{
\method{print}{PrenticeGloeckler.test}(x, ...)
}
\arguments{
  \item{x}{object of class PrenticeGloeckler.test returned by call to PrenticeGloeckler function.}
  \item{...}{ignored}
}

\details{
Prints the coefficient, the estimated log-hazard ratio.
}

\seealso{\code{\link[SurvDisc]{PrenticeGloeckler.test}}}

\author{John Lawrence}

