\name{print.SSDS}
\alias{print.SSDS}
\title{Print Sample Size Discrete Survival Object}
\description{
Print coefficient returned by SampleSizeDiscSurv function.
}
\usage{
\method{print}{SSDS}(x, ...)
}
\arguments{
  \item{x}{object of class SSDS returned by call to  SampleSizeDiscSurv function.}
  \item{...}{ignored}
}

\details{
Prints a few sentences describing the results of the call to the SampleSizeDiscSurv function.
}

\seealso{\code{\link[SurvDisc]{SampleSizeDiscSurv}}}

\author{John Lawrence}

